Move storage listener definition
